move text functions to utils.js
